requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.1

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Text Processing & Splitting
langchain==0.1.0
langchain-text-splitters==0.0.1
tiktoken>=0.5.2

# Vector Database
faiss-cpu==1.12.0
chromadb==0.4.22

# OpenAI Compatible API
openai==1.6.1

# Text Processing
sentence-transformers==2.2.2

# Utilities
numpy>=1.24.3
requests==2.32.3
